requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
pandas
langchain
langchain-openai
langchain-community
upstash-vector
duckdb
langgraph
